requirements.txt
python-kasa>=0.7.0
PyYAML>=6.0.1
aiohttp>=3.9.0
asyncio>=3.4.3
